requirements.txt
werkzeug
flask
numpy
pillow
ISR